requirements.txt
absl-py==1.0.0
alabaster==0.7.12
albumentations==1.1.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
chainer==7.8.1
chainercv==0.13.1
charset-normalizer==2.0.12
cleverhans==4.0.0
cloudpickle==2.0.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.6
docutils==0.17.1
easydict==1.9
entrypoints==0.4
executing==0.8.3
filelock==3.7.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grad-cam==1.3.7
graphviz==0.8.4
grpcio==1.46.1
h5py==3.6.0
idna==3.3
imageio==2.9.0
imagesize==1.3.0
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
keras==2.9.0
Keras-Applications==1.0.8
keras-mxnet==2.2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mnist==0.2.2
msgpack==1.0.3
msgpack-numpy==0.4.7.1
mxnet==1.9.1
nbclassic==0.3.6
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nose==1.3.7
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.21.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.8.0
pycparser==2.21
pyfastcopy==1.0.3
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorchtools==0.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-katex==0.8.6
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-probability==0.16.0
tensorpack==0.11
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.0.0
tifffile==2021.10.12
torch==1.9.0+cu111
torchattacks==3.2.4
torchaudio==0.9.0
-e git+https://github.com/AhmedHussKhalifa/torchdistill_HRS.git@0400fa63baa5e07319172161b625c36d4f4a108b#egg=torchdistill
torchjpeg==0.9.21
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
ttach==0.0.3
typeguard==2.13.3
typing-extensions==3.10.0.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.7.0
